GH-3182: Put the internal API parser classes in their respective parser packages
