+++ jena-arq/src/main/java/org/apache/jena/sparql/lang/sparql_11/ParserSPARQL11.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.sparql.lang;
+package org.apache.jena.sparql.lang.sparql_11;
 
 import java.io.Reader;
 import java.io.StringReader;
@@ -27,38 +27,28 @@ import org.apache.jena.query.QueryException;
 import org.apache.jena.query.QueryParseException;
 import org.apache.jena.query.Syntax;
 import org.apache.jena.shared.JenaException;
-import org.apache.jena.sparql.lang.arq.javacc.ARQParser;
+import org.apache.jena.sparql.lang.SPARQLParser;
 import org.apache.jena.sparql.syntax.Element;
 import org.apache.jena.sparql.syntax.Template;
 
-public class ParserARQ extends SPARQLParser {
-    private interface Action {
-        void exec(ARQParser parser) throws Exception;
-    }
+public class ParserSPARQL11 extends SPARQLParser
+{
+    private interface Action { void exec(SPARQLParser11 parser) throws Exception; }
 
     @Override
     protected Query parse$(final Query query, String queryString) {
-        query.setSyntax(Syntax.syntaxARQ);
+        query.setSyntax(Syntax.syntaxSPARQL_11);
 
-        Action action = new Action() {
-            @Override
-            public void exec(ARQParser parser) throws Exception {
-                parser.QueryUnit();
-            }
-        };
-
+        Action action = (SPARQLParser11 parser) -> parser.QueryUnit();
         perform(query, queryString, action);
         return query;
     }
 
     public static Element parseElement(String string) {
         final Query query = new Query();
-        Action action = new Action() {
-            @Override
-            public void exec(ARQParser parser) throws Exception {
-                Element el = parser.GroupGraphPattern();
-                query.setQueryPattern(el);
-            }
+        Action action = (SPARQLParser11 parser) -> {
+            Element el = parser.GroupGraphPattern();
+            query.setQueryPattern(el);
         };
         perform(query, string, action);
         return query.getQueryPattern();
@@ -66,12 +56,9 @@ public class ParserARQ extends SPARQLParser {
 
     public static Template parseTemplate(String string) {
         final Query query = new Query();
-        Action action = new Action() {
-            @Override
-            public void exec(ARQParser parser) throws Exception {
-                Template t = parser.ConstructTemplate();
-                query.setConstructTemplate(t);
-            }
+        Action action = (SPARQLParser11 parser) -> {
+            Template t = parser.ConstructTemplate();
+            query.setConstructTemplate(t);
         };
         perform(query, string, action);
         return query.getConstructTemplate();
@@ -80,23 +67,21 @@ public class ParserARQ extends SPARQLParser {
     // All throwable handling.
     private static void perform(Query query, String string, Action action) {
         Reader in = new StringReader(string);
-        ARQParser parser = new ARQParser(in);
+        SPARQLParser11 parser = new SPARQLParser11(in);
 
         try {
             query.setStrict(true);
             parser.setQuery(query);
             action.exec(parser);
-        } catch (org.apache.jena.sparql.lang.arq.javacc.ParseException ex) {
+        } catch (org.apache.jena.sparql.lang.sparql_11.ParseException ex) {
             throw new QueryParseException(ex.getMessage(), ex.currentToken.beginLine, ex.currentToken.beginColumn);
-        } catch (org.apache.jena.sparql.lang.arq.javacc.TokenMgrError tErr) {
+        } catch (org.apache.jena.sparql.lang.sparql_11.TokenMgrError tErr) {
             // Last valid token : not the same as token error message - but this
             // should not happen
             int col = parser.token.endColumn;
             int line = parser.token.endLine;
             throw new QueryParseException(tErr.getMessage(), line, col);
-        }
-
-        catch (QueryException ex) {
+        } catch (QueryException ex) {
             throw ex;
         } catch (JenaException ex) {
             throw new QueryException(ex.getMessage(), ex);
