+++ jena-arq/src/main/java/org/apache/jena/sparql/lang/sparql_12/ParserSPARQL12.java
@@ -16,30 +16,63 @@
  * limitations under the License.
  */
 
-package org.apache.jena.sparql.lang;
+package org.apache.jena.sparql.lang.sparql_12;
 
 import java.io.Reader;
+import java.io.StringReader;
 
 import org.apache.jena.atlas.logging.Log;
+import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryException;
 import org.apache.jena.query.QueryParseException;
+import org.apache.jena.query.Syntax;
 import org.apache.jena.shared.JenaException;
-import org.apache.jena.sparql.core.Prologue;
+import org.apache.jena.sparql.lang.SPARQLParser;
 import org.apache.jena.sparql.lang.sparql_12.javacc.SPARQLParser12;
-import org.apache.jena.sparql.modify.UpdateSink;
-import org.apache.jena.update.UpdateException;
+import org.apache.jena.sparql.syntax.Element;
+import org.apache.jena.sparql.syntax.Template;
 
-public class ParserSPARQL12Update extends UpdateParser
+public class ParserSPARQL12 extends SPARQLParser
 {
-    public ParserSPARQL12Update() {}
+    private interface Action { void exec(SPARQLParser12 parser) throws Exception; }
 
     @Override
-    protected void executeParse(UpdateSink sink, Prologue prologue, Reader r) {
-        SPARQLParser12 parser = null;
+    protected Query parse$(final Query query, String queryString) {
+        query.setSyntax(Syntax.syntaxSPARQL_12);
+        Action action = (SPARQLParser12 parser) -> parser.QueryUnit();
+        perform(query, queryString, action);
+        return query;
+    }
+
+    public static Element parseElement(String string) {
+        final Query query = new Query();
+        Action action = (SPARQLParser12 parser) -> {
+            Element el = parser.GroupGraphPattern();
+            query.setQueryPattern(el);
+        };
+        perform(query, string, action);
+        return query.getQueryPattern();
+    }
+
+    public static Template parseTemplate(String string) {
+        final Query query = new Query();
+        Action action = (SPARQLParser12 parser) -> {
+            Template t = parser.ConstructTemplate();
+            query.setConstructTemplate(t);
+        };
+        perform(query, string, action);
+        return query.getConstructTemplate();
+    }
+
+    // All throwable handling.
+    private static void perform(Query query, String string, Action action) {
+        Reader in = new StringReader(string);
+        SPARQLParser12 parser = new SPARQLParser12(in);
+
         try {
-            parser = new SPARQLParser12(r);
-            parser.setUpdate(prologue, sink);
-            parser.UpdateUnit();
+            query.setStrict(true);
+            parser.setQuery(query);
+            action.exec(parser);
         } catch (org.apache.jena.sparql.lang.sparql_12.javacc.ParseException ex) {
             throw new QueryParseException(ex.getMessage(), ex.currentToken.beginLine, ex.currentToken.beginColumn);
         } catch (org.apache.jena.sparql.lang.sparql_12.javacc.TokenMgrError tErr) {
@@ -48,7 +81,7 @@ public class ParserSPARQL12Update extends UpdateParser
             int col = parser.token.endColumn;
             int line = parser.token.endLine;
             throw new QueryParseException(tErr.getMessage(), line, col);
-        } catch (UpdateException ex) {
+        } catch (QueryException ex) {
             throw ex;
         } catch (JenaException ex) {
             throw new QueryException(ex.getMessage(), ex);
@@ -56,7 +89,7 @@ public class ParserSPARQL12Update extends UpdateParser
             // The token stream can throw errors.
             throw new QueryParseException(err.getMessage(), err, -1, -1);
         } catch (Throwable th) {
-            Log.error(this, "Unexpected throwable: ", th);
+            Log.warn(ParserSPARQL12.class, "Unexpected throwable: ", th);
             throw new QueryException(th.getMessage(), th);
         }
     }
