+++ jena-arq/src/main/java/org/apache/jena/sparql/lang/sparql_12/ParserSPARQL12Update.java
@@ -16,62 +16,31 @@
  * limitations under the License.
  */
 
-package org.apache.jena.sparql.lang;
+package org.apache.jena.sparql.lang.sparql_12;
 
 import java.io.Reader;
-import java.io.StringReader;
 
 import org.apache.jena.atlas.logging.Log;
-import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryException;
 import org.apache.jena.query.QueryParseException;
-import org.apache.jena.query.Syntax;
 import org.apache.jena.shared.JenaException;
+import org.apache.jena.sparql.core.Prologue;
+import org.apache.jena.sparql.lang.UpdateParser;
 import org.apache.jena.sparql.lang.sparql_12.javacc.SPARQLParser12;
-import org.apache.jena.sparql.syntax.Element;
-import org.apache.jena.sparql.syntax.Template;
+import org.apache.jena.sparql.modify.UpdateSink;
+import org.apache.jena.update.UpdateException;
 
-public class ParserSPARQL12 extends SPARQLParser
+public class ParserSPARQL12Update extends UpdateParser
 {
-    private interface Action { void exec(SPARQLParser12 parser) throws Exception; }
+    public ParserSPARQL12Update() {}
 
     @Override
-    protected Query parse$(final Query query, String queryString) {
-        query.setSyntax(Syntax.syntaxSPARQL_12);
-        Action action = (SPARQLParser12 parser) -> parser.QueryUnit();
-        perform(query, queryString, action);
-        return query;
-    }
-
-    public static Element parseElement(String string) {
-        final Query query = new Query();
-        Action action = (SPARQLParser12 parser) -> {
-            Element el = parser.GroupGraphPattern();
-            query.setQueryPattern(el);
-        };
-        perform(query, string, action);
-        return query.getQueryPattern();
-    }
-
-    public static Template parseTemplate(String string) {
-        final Query query = new Query();
-        Action action = (SPARQLParser12 parser) -> {
-            Template t = parser.ConstructTemplate();
-            query.setConstructTemplate(t);
-        };
-        perform(query, string, action);
-        return query.getConstructTemplate();
-    }
-
-    // All throwable handling.
-    private static void perform(Query query, String string, Action action) {
-        Reader in = new StringReader(string);
-        SPARQLParser12 parser = new SPARQLParser12(in);
-
+    protected void executeParse(UpdateSink sink, Prologue prologue, Reader r) {
+        SPARQLParser12 parser = null;
         try {
-            query.setStrict(true);
-            parser.setQuery(query);
-            action.exec(parser);
+            parser = new SPARQLParser12(r);
+            parser.setUpdate(prologue, sink);
+            parser.UpdateUnit();
         } catch (org.apache.jena.sparql.lang.sparql_12.javacc.ParseException ex) {
             throw new QueryParseException(ex.getMessage(), ex.currentToken.beginLine, ex.currentToken.beginColumn);
         } catch (org.apache.jena.sparql.lang.sparql_12.javacc.TokenMgrError tErr) {
@@ -80,7 +49,7 @@ public class ParserSPARQL12 extends SPARQLParser
             int col = parser.token.endColumn;
             int line = parser.token.endLine;
             throw new QueryParseException(tErr.getMessage(), line, col);
-        } catch (QueryException ex) {
+        } catch (UpdateException ex) {
             throw ex;
         } catch (JenaException ex) {
             throw new QueryException(ex.getMessage(), ex);
@@ -88,7 +57,7 @@ public class ParserSPARQL12 extends SPARQLParser
             // The token stream can throw errors.
             throw new QueryParseException(err.getMessage(), err, -1, -1);
         } catch (Throwable th) {
-            Log.warn(ParserSPARQL12.class, "Unexpected throwable: ", th);
+            Log.error(this, "Unexpected throwable: ", th);
             throw new QueryException(th.getMessage(), th);
         }
     }
